requirements.txt
streamlit==1.46.0
groq==0.28.0
python-dotenv==1.1.1
langchain==0.0.350
pydantic==2.11.7
rich==14.0.0
typing-extensions==4.8.0